requirement.txt
This file lists all the Python libraries needed to run the project.
To install them all, run: pip install -r requirements.txt
streamlit
openai-whisper
srt